+++ packages/vendure-plugin-stripe-subscription/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.4.0 (2024-12-02)
+
+- Made `isSubscription()` async and passed an instance of `Injector`, so that consumers can fetch additional relations inside `isSubscription()`;
+- `subscriptionHash` custom field was removed: The plugin doesn't need an order line per subscription per se.
+
 # 2.3.1 (2023-12-26)
 
 - Correctly calculate line price sum when multiple subscriptions are returned